test: keep formula v2 helper out of production package
