+++ cmd/gc/cmd_sling_routevars_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/gastownhall/gascity/internal/beads"
 	"github.com/gastownhall/gascity/internal/config"
 	"github.com/gastownhall/gascity/internal/formula"
+	"github.com/gastownhall/gascity/internal/formulatest"
 	"github.com/gastownhall/gascity/internal/molecule"
 	"github.com/gastownhall/gascity/internal/sling"
 )
@@ -146,7 +147,7 @@ type = "task"
 	}
 
 	// Enable graph workflow features.
-	formula.EnableV2ForTest(t)
+	formulatest.EnableV2ForTest(t)
 	prevGraphApply := molecule.IsGraphApplyEnabled()
 	molecule.SetGraphApplyEnabled(true)
 	t.Cleanup(func() {
