+++ internal/formula/testhelper_test.go
@@ -1,0 +1,10 @@
+package formula
+
+import "testing"
+
+func enableV2ForTest(tb testing.TB) {
+	tb.Helper()
+	prev := IsFormulaV2Enabled()
+	SetFormulaV2Enabled(true)
+	tb.Cleanup(func() { SetFormulaV2Enabled(prev) })
+}
